Extract parser, decouple parser and renderer using document, split document into data container and api
